+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * refactoring $onMount
 * optional deep checking for passed props: prop={value} prop|deep={value}
 * keep-alive
+* malina:self
 
 ## 0.6.x